Removed api module and replace with module with functions that can be called directly without instantiating a class.
cli imports and uses modules. Creates pipeline and specifcation
objects and store in command context for use in other commands.

Removed a few unused dependencies.
