+++ digital_land/cli.py
@@ -1,115 +1,62 @@
-import functools
 import logging
 import sys
-from collections import defaultdict
-
 import click
 
-from digital_land.api import DigitalLandApi
+from collections import defaultdict
 
-# pyright: reportGeneralTypeIssues=false
-API: DigitalLandApi = None
+from digital_land.commands import (
+    fetch,
+    collect,
+    collection_list_resources,
+    collection_pipeline_makerules,
+    dataset_dump,
+    dataset_dump_flattened,
+    collection_save_csv,
+    convert,
+    dataset_create,
+    pipeline_run,
+    collection_add_source,
+    build_datasette,
+    expectations,
+)
 
-
-# Custom decorators for common command arguments
-def input_output_path(f):
-    arguments = [
-        click.argument("input-path", type=click.Path(exists=True)),
-        click.argument("output-path", type=click.Path(), default=""),
-    ]
-    return functools.reduce(lambda x, arg: arg(x), reversed(arguments), f)
-
-
-def dataset(f):
-    return click.option("--dataset", "-n", type=click.STRING)(f)
-
-
-def pipeline_dir(f):
-    return click.option("--pipeline-dir", "-p", type=click.Path(), default="pipeline/")(
-        f
-    )
-
-
-def collection_dir(f):
-    return click.option(
-        "--collection-dir",
-        "-c",
-        type=click.Path(exists=True),
-        default="collection/",
-    )(f)
-
-
-def specification_dir(f):
-    return click.option(
-        "--specification-dir",
-        "-s",
-        type=click.Path(exists=True),
-        default="specification/",
-    )(f)
-
-
-def issue_dir(f):
-    return click.option(
-        "--issue-dir", "-i", type=click.Path(exists=True), default="issue/"
-    )(f)
-
-
-def column_field_dir(f):
-    return click.option(
-        "--column-field-dir",
-        type=click.Path(exists=True),
-        default="var/column-field/",
-    )(f)
-
-
-def dataset_resource_dir(f):
-    return click.option(
-        "--dataset-resource-dir",
-        type=click.Path(exists=True),
-        default="var/dataset-resource/",
-    )(f)
-
-
-def endpoint_path(f):
-    return click.option(
-        "--endpoint-path",
-        type=click.Path(exists=True),
-        default="collection/endpoint.csv",
-    )(f)
-
-
-def source_path(f):
-    return click.option(
-        "--source-path",
-        type=click.Path(exists=True),
-        default="collection/source.csv",
-    )(f)
-
-
-def organisation_path(f):
-    return click.option(
-        "--organisation-path",
-        "-o",
-        type=click.Path(exists=True),
-        default="var/cache/organisation.csv",
-    )(f)
+from digital_land.command_arguments import (
+    collection_dir,
+    organisation_path,
+    input_output_path,
+    issue_dir,
+    dataset_resource_dir,
+    column_field_dir,
+)
 
 
 @click.group()
-@click.option("-d", "--debug/--no-debug", default=False)
-@dataset
-@pipeline_dir
-@specification_dir
-def cli(debug, dataset, pipeline_dir, specification_dir):
-    global API
-    API = DigitalLandApi(debug, dataset, pipeline_dir, specification_dir)
+@click.option("-d", "--debug/--no-debug", type=click.BOOL, default=False)
+@click.option("--dataset", "-n", type=click.STRING)
+@click.option("--pipeline-dir", "-p", type=click.Path(), default="pipeline/")
+@click.option(
+    "--specification-dir", "-s", type=click.Path(exists=True), default="specification/"
+)
+@click.pass_context
+def cli(ctx, debug, dataset, pipeline_dir, specification_dir):
+
+    ctx.ensure_object(dict)
+
+    from digital_land.pipeline import Pipeline
+    from digital_land.specification import Specification
+
+    ctx.obj["PIPELINE"] = Pipeline(pipeline_dir, dataset)
+    ctx.obj["SPECIFICATION"] = Specification(specification_dir)
+    ctx.obj["DATASET"] = dataset
+    ctx.obj["DEBUG"] = debug
 
 
 @cli.command("fetch")
 @click.argument("url")
-def fetch_cmd(url):
+@click.pass_context
+def fetch_cmd(ctx, url):
     """fetch resource from a single endpoint"""
-    return API.fetch_cmd(url)
+    return fetch(url, ctx.obj["PIPELINE"])
 
 
 @cli.command("collect")
@@ -119,19 +66,19 @@ def fetch_cmd(url):
     default="collection/endpoint.csv",
 )
 @collection_dir
-def collect_cmd(endpoint_path, collection_dir):
+@click.pass_context
+def collect_cmd(ctx, endpoint_path, collection_dir):
     """fetch resources from collection endpoints"""
-    return API.collect_cmd(endpoint_path, collection_dir)
+    return collect(endpoint_path, collection_dir, ctx.obj["PIPELINE"])
 
 
 #
 #  collection commands
-#  TBD: make sub commands
 #
 @cli.command("collection-list-resources", short_help="list resources for a pipeline")
 @collection_dir
 def collection_list_resources_cmd(collection_dir):
-    return API.collection_list_resources_cmd(collection_dir)
+    return collection_list_resources(collection_dir)
 
 
 @cli.command(
@@ -140,13 +87,13 @@ def collection_list_resources_cmd(collection_dir):
 )
 @collection_dir
 def collection_pipeline_makerules_cmd(collection_dir):
-    return API.collection_pipeline_makerules_cmd(collection_dir)
+    return collection_pipeline_makerules(collection_dir)
 
 
 @cli.command("collection-save-csv", short_help="save collection as CSV package")
 @collection_dir
 def collection_save_csv_cmd(collection_dir):
-    return API.collection_save_csv_cmd(collection_dir)
+    return collection_save_csv(collection_dir)
 
 
 #
@@ -155,21 +102,27 @@ def collection_save_csv_cmd(collection_dir):
 @cli.command("convert", short_help="convert a resource to CSV")
 @input_output_path
 def convert_cmd(input_path, output_path):
-    return API.convert_cmd(input_path, output_path)
+    return convert(input_path, output_path)
 
 
 @cli.command("dataset-create", short_help="create a dataset from processed resources")
-@organisation_path
 @click.option("--output-path", type=click.Path(), default=None, help="sqlite3 path")
 @click.argument("input-paths", nargs=-1, type=click.Path(exists=True))
-def dataset_create_cmd(input_paths, output_path, organisation_path):
-    return API.dataset_create_cmd(input_paths, output_path, organisation_path)
+@organisation_path
+@click.pass_context
+def dataset_create_cmd(ctx, input_paths, output_path, organisation_path):
+    pipeline = ctx.obj["PIPELINE"]
+    specification = ctx.obj["SPECIFCATION"]
+    dataset = ctx.obj["DATASET"]
+    return dataset_create(
+        input_paths, output_path, organisation_path, pipeline, specification, dataset
+    )
 
 
 @cli.command("dataset-entries", short_help="dump dataset entries as csv")
 @input_output_path
 def dataset_dump_cmd(input_path, output_path):
-    API.dataset_dump_cmd(input_path, output_path)
+    dataset_dump(input_path, output_path)
 
 
 @cli.command(
@@ -177,22 +130,27 @@ def dataset_dump_cmd(input_path, output_path):
     short_help="dump dataset entries as csv with additional top-level `entity.json` fields",
 )
 @input_output_path
-def dataset_dump_flattened_cmd(input_path, output_path):
-    API.dataset_dump_flattened_cmd(csv_path=input_path, flattened_dir=output_path)
+@click.pass_context
+def dataset_dump_flattened_cmd(ctx, input_path, output_path):
+    specification = ctx["specfication"]
+    dataset = ctx["dataset"]
+    dataset_dump_flattened(input_path, output_path, specification, dataset)
 
 
 @cli.command("pipeline", short_help="process a resource")
-@input_output_path
-@issue_dir
-@column_field_dir
-@dataset_resource_dir
-@organisation_path
 @click.option("--save-harmonised", is_flag=True)
 @click.option("--endpoints", help="list of endpoint hashes", default="")
 @click.option("--organisations", help="list of organisations", default="")
 @click.option("--entry-date", help="default entry-date value", default="")
 @click.option("--custom-temp-dir", help="default temporary directory", default=None)
-def pipeline_cmd_API(
+@input_output_path
+@issue_dir
+@column_field_dir
+@dataset_resource_dir
+@organisation_path
+@click.pass_context
+def pipeline_command(
+    ctx,
     input_path,
     output_path,
     issue_dir,
@@ -205,9 +163,17 @@ def pipeline_cmd_API(
     entry_date,
     custom_temp_dir,
 ):
+    dataset = ctx.obj["DATASET"]
+    pipeline = ctx.obj["PIPELINE"]
+    specification = ctx.obj["SPECIFCATION"]
+
     endpoints = endpoints.split()
     organisations = organisations.split()
-    return API.pipeline_cmd(
+
+    return pipeline_run(
+        dataset,
+        pipeline,
+        specification,
         input_path,
         output_path,
         issue_dir=issue_dir,
@@ -230,7 +196,6 @@ def pipeline_cmd_API(
     short_help="Add a new source and endpoint to a collection",
     context_settings=dict(ignore_unknown_options=True, allow_extra_args=True),
 )
-@click.pass_context
 @click.argument("collection", type=click.STRING)
 @click.argument("endpoint-url", type=click.STRING)
 @collection_dir
@@ -247,9 +212,7 @@ def collection_add_source_cmd(ctx, collection, endpoint_url, collection_dir):
         str,
         {ctx.args[i]: ctx.args[i + 1] for i in range(0, len(ctx.args), 2)},
     )
-    return API.collection_add_source_cmd(
-        entry, collection, endpoint_url, collection_dir
-    )
+    return collection_add_source(entry, collection, endpoint_url, collection_dir)
 
 
 @cli.command("build-datasette", short_help="build docker image for datasette")
@@ -257,8 +220,8 @@ def collection_add_source_cmd(ctx, collection, endpoint_url, collection_dir):
 @click.option("--data-dir", default="./var/cache")
 @click.option("--ext", default="sqlite3")
 @click.option("--options", default=None)
-def build_datasette(tag, data_dir, ext, options):
-    return API.build_datasette(tag, data_dir, ext, options)
+def build_datasette_cmd(tag, data_dir, ext, options):
+    return build_datasette(tag, data_dir, ext, options)
 
 
 @cli.command("expectations", short_help="runs data quality expectations")
@@ -268,4 +231,4 @@ def build_datasette(tag, data_dir, ext, options):
 )
 @click.option("--data-quality-yaml", help="path to expectations yaml", required=True)
 def call_expectations(results_path, sqlite_dataset_path, data_quality_yaml):
-    return API.expectation_cmd(results_path, sqlite_dataset_path, data_quality_yaml)
+    return expectations(results_path, sqlite_dataset_path, data_quality_yaml)
