+++ docker-compose.yml
@@ -1,0 +1,9 @@
+services:
+
+  ubuntu:
+    build: .
+    volumes:
+      - .:/src
+    stdin_open: true
+    tty: true
+
