+++ tests/integration/test_package_dataset.py
@@ -1,11 +1,12 @@
 import json
 import os.path
-from csv import DictReader
-
 import pytest
 
-from digital_land.api import DigitalLandApi
+from csv import DictReader
+
+from digital_land.commands import dataset_create, dataset_dump, dataset_dump_flattened
 from digital_land.specification import specification_path
+from digital_land.specification import Specification
 
 
 @pytest.mark.skip("data needs updating")
@@ -45,15 +46,8 @@ def test_package_dataset(
     sqlite_path = output_dir.joinpath(f"{dataset_name}.sqlite3")
     csv_path = output_dir.joinpath(f"{dataset_name}.csv")
 
-    # Call
-    api = DigitalLandApi(
-        debug=False,
-        dataset=dataset_name,
-        pipeline_dir=pipeline_dir,
-        specification_dir=str(specification_path),
-    )
-    api.dataset_create_cmd(input_paths, sqlite_path, organisation_path)
-    api.dataset_dump_cmd(sqlite_path, csv_path)
+    dataset_create(input_paths, sqlite_path, organisation_path)
+    dataset_dump(sqlite_path, csv_path)
 
     # Assert
     with csv_path.open() as actual, expected_csv_result.open() as expected:
@@ -67,7 +61,7 @@ def test_package_dataset(
     "dataset_name",
     ["listed-building-grade"],
 )
-def test_package_dataset_hoisted(
+def test_package_flattened_dataset(
     # Parametrize args
     dataset_name,
     # Runtime filesystem dependencies generated by previous steps
@@ -87,14 +81,9 @@ def test_package_dataset_hoisted(
     flattened_csv_path = flattened_output_dir.joinpath(f"{dataset_name}.csv")
     flattened_json_path = flattened_output_dir.joinpath(f"{dataset_name}.json")
 
-    # Call
-    api = DigitalLandApi(
-        debug=False,
-        dataset=dataset_name,
-        pipeline_dir=pipeline_dir,
-        specification_dir=str(specification_path),
-    )
-    api.dataset_dump_flattened_cmd(csv_path, flattened_output_dir)
+    specification = Specification(specification_path)
+
+    dataset_dump_flattened(csv_path, flattened_output_dir, specification, dataset_name)
 
     with flattened_csv_path.open() as actual, expected_flattened_csv_result.open() as expected:
         actual_dict_reader = DictReader(actual)
